+++ frontend/src/module/book/components/PagesNavigation.tsx
