+++ frontend/src/module/book/components/SidePanel/ColorPanel.tsx
