+++ frontend/src/module/book/components/SidePanel/GraphicsPanel.tsx
