+++ frontend/src/module/book/components/SidePanel/PagesPanel.tsx
