+++ frontend/src/module/book/components/SidePanel/SidePanel.tsx
