+++ frontend/src/module/book/components/SidePanel/VerticalSeparator.tsx
